Add JWT Authentication and Authorization.
A JWT may be obtained by POSTing the following credentials to '/login'. If successful, the reponse will contain the JWT in the Authorization header.

"email": "myEmail",
"password": "myPassword".

This token must be added to the Authorization header of all requests that aren't public (see WebSecurity.AUTH_WHITELIST for public endpoints).